fix type error.
* iseq.c (iseq_iterate_children): should use cast to `int`.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@66005 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ iseq.c
@@ -2214,7 +2214,7 @@ iseq_iterate_children(const rb_iseq_t *iseq, void (*iter_func)(const rb_iseq_t *
         i += len;
     }
 
-    return RHASH_SIZE(all_children);
+    return (int)RHASH_SIZE(all_children);
 }
 
 static void
